+++ src/packages/piral-cli/src/apps/debug-piral.ts
@@ -3,7 +3,6 @@ import { readKrasConfig, krasrc, buildKrasWithCli, defaultConfig } from 'kras';
 import {
   retrievePiletsInfo,
   retrievePiralRoot,
-  clearCache,
   setStandardEnvs,
   openBrowser,
   checkCliCompatibility,
@@ -12,6 +11,7 @@ import {
   logInfo,
   patchModules,
   setupBundler,
+  removeDirectory,
 } from '../common';
 
 export interface DebugPiralOptions {
@@ -60,6 +60,7 @@ export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOpt
   } = options;
   const entryFiles = await retrievePiralRoot(baseDir, entry);
   const { externals, name, root, ignored } = await retrievePiletsInfo(entryFiles);
+  const cache = resolve(root, cacheDir);
 
   await checkCliCompatibility(root);
 
@@ -86,12 +87,12 @@ export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOpt
   }
 
   if (fresh) {
-    await clearCache(root, cacheDir);
+    await removeDirectory(cache);
   }
 
   if (optimizeModules) {
     logInfo('Preparing modules ...');
-    await patchModules(root, cacheDir, ignored);
+    await patchModules(root, cache, ignored);
   }
 
   setStandardEnvs({
@@ -106,7 +107,7 @@ export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOpt
     config: {
       publicUrl,
       logLevel,
-      cacheDir,
+      cacheDir: cache,
       scopeHoist,
       hmr,
       autoInstall,