+++ src/packages/piral-cli/src/common/info.ts
@@ -16,6 +16,7 @@ export const nodeVersion = process.version.substr(1);
 export const cliVersion = info.version;
 export const compatVersion = findCompatVersion(cliVersion);
 export const isWindows = process.platform === 'win32';
+export const defaultCacheDir = '.cache';
 export const pathSeparator = isWindows ? ';' : ':';
 export const defaultRegistry = 'https://registry.npmjs.org/';
 export const debugPiletApi = '/$pilet-api';